+++ structures/tests/testdata/__init__.py
@@ -286,6 +286,7 @@ def load_notification_entities(owner: Owner):
         del x['owner_corporation_id']
         Structure.objects.create(**x)
     
+    timestamp_start = now() - timedelta(hours=2)
     for notification in entities_testdata['Notification']:
         notification_type = \
             Notification.get_matching_notification_type(
@@ -301,16 +302,13 @@ def load_notification_entities(owner: Owner):
                 if 'text' in notification else None
             is_read = notification['is_read'] \
                 if 'is_read' in notification else None
+            timestamp_start = timestamp_start + timedelta(minutes=5)
             obj = Notification.objects.update_or_create(
                 notification_id=notification['notification_id'],
                 owner=owner,
                 defaults={
                     'sender': sender,
-                    'timestamp': now() - timedelta(
-                        hours=randrange(3), 
-                        minutes=randrange(60), 
-                        seconds=randrange(60)
-                    ),
+                    'timestamp': timestamp_start,
                     'notification_type': notification_type,
                     'text': text,
                     'is_read': is_read,